[FEATURE] Use typo3/tailor to publish to TER.
This publishes a new version to TER if a new tag of the form `v1.2.3` is found.

Mind that version numbers in `ext_emconf.php` and `./Documentation/Settings.cfg` must
reflect the new version.

In order to facilitate that, a composer command "set-version" is introduced as shortcut
to typo3/tailor.

These steps will release a new version:
* `ddev composer set-version 1.3.0`
* Commit the changes. Your commit message will be used as release notes for TER, too.
* `git tag v1.3.0` (notice the 'v' in front of the release tag)
* git push && git push --tags

This will make the new version available to Packagist (via Github Webhook) and TER (via API)
and trigger rerendering of the documentation on docs.typo3.org (via Github Webhook).